Conflicts:
	flutter/test-app-local-dependency/android/app/build.gradle
	flutter/test-app-local-dependency/pubspec.lock
	flutter/test-app-local-dependency/pubspec.yaml
	flutter/test-app-pub/pubspec.lock

+++ flutter/test-app-local-dependency/lib/saf_tab.dart
@@ -164,7 +164,7 @@ class SafTab {
       return;
     }
 
-    int timeInMilliseconds = statistics.getTime();
+    double timeInMilliseconds = statistics.getTime();
     int totalVideoDuration = 9000;
 
     int completePercentage = (timeInMilliseconds * 100) ~/ totalVideoDuration;
